requirements.txt
Django~=4.1.5
requests~=2.28.2
djangorestframework~=3.14.0